Fix source links for gems hosted on Github
The gems lucene_query_parser and sentry-raven are set in the Gemfile to be installed from git repositories instead of from rubygems.org. When I tried to run bundle install, I got error messages saying that Github couldn't be reached and I was unable to install those two gems. So, I changed the two links to Github to the HTTPS clone URL for the two repos. Also, for sentry-raven, I changed the key from :github to :git. Then I was able to bundle and run the application.